+++ src/test/unit/snykOss/services/ossService.test.ts
@@ -33,6 +33,8 @@ suite('OssService', () => {
       logger,
       {
         getAdditionalCliParameters: () => '',
+        getCustomCliPath: () => undefined,
+        isAutomaticDependencyManagementEnabled: () => true,
       } as IConfiguration,
       {} as IWebViewProvider<OssIssueCommandArg>,
       {